Update tests for new chain break methods
