+++ dimod/embedding/__init__.py
@@ -14,7 +14,6 @@
 #
 # ================================================================================================
 
-from dimod.embedding.array import *
 from dimod.embedding.chain_breaks import *
 from dimod.embedding.transforms import *
 from dimod.embedding.utils import *
